Emotion-Detection-2/EfficientNet/requirements.txt
torch
torchvision
efficientnet_pytorch
tqdm
scikit-learn
matplotlib
seaborn
Pillow
